# docs/swagger.yaml
basePath: /
definitions:
  models.Gender:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.PatchGender:
    properties:
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for enriching names
  title: Name Enricher API
  version: "1.0"
paths:
  /genders:
    get:
      consumes:
      - application/json
      description: Get a list of genders with optional filtering
      parameters:
      - description: Gender ID
        in: query
        name: id
        type: integer
      - description: Gender name
        in: query
        name: name
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved gender list
          schema:
            items:
              $ref: '#/definitions/models.Gender'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List genders
      tags:
      - genders
    post:
      consumes:
      - application/json
      description: Create a new gender entry in the database
      parameters:
      - description: Gender object with name field
        in: body
        name: gender
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created gender
          schema:
            $ref: '#/definitions/models.Gender'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new gender
      tags:
      - genders
  /genders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a gender by its ID
      parameters:
      - description: Gender ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted gender
          schema:
            $ref: '#/definitions/models.Gender'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a gender
      tags:
      - genders
    get:
      consumes:
      - application/json
      description: Get a single gender by its ID
      parameters:
      - description: Gender ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved gender
          schema:
            $ref: '#/definitions/models.Gender'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Gender not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a gender by ID
      tags:
      - genders
    put:
      consumes:
      - application/json
      description: Update an existing gender by ID
      parameters:
      - description: Gender ID
        in: path
        name: id
        required: true
        type: integer
      - description: Gender update data
        in: body
        name: gender
        required: true
        schema:
          $ref: '#/definitions/models.PatchGender'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated gender
          schema:
            $ref: '#/definitions/models.Gender'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a gender
      tags:
      - genders
...
swagger: "2.0"